react native tests and frontend

+++ backend/resolvers.js
@@ -1,6 +1,5 @@
 const { PubSub, AuthenticationError, UserInputError } = require('apollo-server')
 const Location = require('./models/Location')
-const Play = require('./models/Play')
 const Point = require('./models/Point')
 const Round = require('./models/Round')
 const User = require('./models/User')
@@ -40,18 +39,27 @@ const resolvers = {
     },
   }),
   Query: {
-    allUsers: async () => {
+    allUsers: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
+      }
       const users = await User
         .find({})
         .populate('friends')
       return users
     },
-    allLocations: async () => {
+    allLocations: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
+      }
       const locations = await Location
         .find({})
       return locations
     },
     allRounds: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
+      }
       if (args.id) {
         let round = await Round
           .findById(args.id)
@@ -77,26 +85,12 @@ const resolvers = {
       if (args.location) {
         rounds = rounds.filter(round => round.location.name = args.location)
       }
-      console.log('all rounds', rounds)
       return rounds
     },
-    allPlays: async (root, args) => {
-      let plays = await Play
-        .find({})
-        .populate('round')
-        .populate('user')
-      if (args.roundId) {
-        plays = plays.filter(play => play.round.id === args.roundId)
+    allPoints: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
       }
-      if (args.playNumber) {
-        plays = plays.filter(play => play.number === args.playNumber)
-      }
-      if (args.username) {
-        plays = plays.filter(play => play.user.username === args.username)
-      }
-      return plays
-    },
-    allPoints: async (root, args) => {
       let points = []
       if (args.roundId) {
         points = await Point
@@ -106,19 +100,21 @@ const resolvers = {
         if (args.userId) {
           points = points.filter(point => point.user._id === args.userId)
         }
-        if (args.trackNumber) {
-          points = points.filter(point => point.trackNumber === args.trackNumber)
+        if (args.trackIndex) {
+          points = points.filter(point => point.trackIndex === args.trackIndex)
         }
       }
       return points
     },
-    allFriends: async (root, args) => {
+    allFriends: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
+      }
       let user
       if (args.username) {
         user = await User
           .findOne({ username: args.username })
           .populate('friends')
-        console.log('user friends', user)
       } else if (args.userId) {
         user = await User
           .findById(args.userId)
@@ -135,7 +131,6 @@ const resolvers = {
       }
       const user = await User.findById(currentUser._id)
         .populate('friends')
-      console.log('user', user)
       return user
     },
   },
@@ -143,12 +138,12 @@ const resolvers = {
     login: async (root, args) => {
       const user = await User.findOne({ username: args.username })
       if (!user) {
-        throw new UserInputError('Username not found');
+        throw new UserInputError('Wrong username or password');
       }
 
       const valid = await bcrypt.compare(args.password, user.passwordHash);
       if (!valid) {
-        throw new UserInputError('Wrong password');
+        throw new UserInputError('Wrong username or password');
       }
       const userForToken = {
         username: user.username,
@@ -156,8 +151,10 @@ const resolvers = {
       }
       return { token: jwt.sign(userForToken, JWT_SECRET), username: user.username }
     },
-    addUser: async (root, args) => {
-      console.log('create user', args)
+    addUser: async (root, args, { currentUser }) => {
+      if (!currentUser || !currentUser.admin) {
+        throw new UserInputError('invalid token');
+      }
       if (args.password.length < 3) {
         throw new UserInputError('password should be longer than 3 letters')
       }
@@ -185,10 +182,6 @@ const resolvers = {
       if (!currentUser) {
         throw new UserInputError('invalid token');
       }
-      console.log('make friends with', args.username)
-      if (!currentUser) {
-        throw new AuthenticationError('Invalid authentication token')
-      }
       let friend = await User
         .findOne({ username: args.username })
       if (!friend) {
@@ -196,8 +189,6 @@ const resolvers = {
           invalidArgs: args.username
         })
       }
-      console.log('friend', friend)
-      console.log('index of friend', currentUser.friends.indexOf(friend._id))
       if (currentUser.friends.indexOf(friend._id) > -1) {
         throw new UserInputError('you are already friends', {
           invalidArgs: args.username
@@ -235,7 +226,6 @@ const resolvers = {
       return location
     },
     addRound: async (root, args, { currentUser }) => {
-      console.log('users', args.userIds, 'location', args.locationId)
       if (!currentUser) {
         throw new UserInputError('invalid token');
       }
@@ -251,43 +241,45 @@ const resolvers = {
           .populate('location')
           .populate('users')
       } catch (error) {
-        console.log('error', error.message)
         throw new UserInputError(error.message, {
           invalidArgs: args
         })
       }
     },
-    addPlay: async (root, args, { currentUser }) => {
+    addPoint: async (root, args, { currentUser }) => {
       if (!currentUser) {
         throw new UserInputError('invalid token');
       }
-      let play = await Play
+      let point = await Point
         .findOne({
-          playNumber: args.playNumber,
+          trackIndex: args.trackIndex,
           round: args.roundId,
+          user: args.userId,
         })
-      if (!play) {
-        play = new Play({
-          playNumber: args.playNumber,
+      if (!point) {
+        point = new Point({
+          trackIndex: args.trackIndex,
           round: args.roundId,
+          user: args.userId,
+          points: args.points
         })
         try {
-          await play.save()
+          await point.save()
         } catch (error) {
           throw new UserInputError(error.message, {
             invalidArgs: args
           })
         }
       }
-      return await play
+      return await Point.findById(point.id)
         .populate('round')
+        .populate('user')
     },
 
     addCachedPoints: async (root, args, { currentUser }) => {
       if (!currentUser) {
         throw new UserInputError('invalid token');
       }
-      console.log('add point args', args)
       const roundId = args.roundId
       const round = await Round.findById(roundId)
       const userIds = args.userIds
@@ -310,7 +302,6 @@ const resolvers = {
         const user = await User.findById(userId)
         const trackIndex = trackIndexes[i]
         const point = points[i]
-        console.log('check point 1', userId, roundId)
         const updatedPoint = await Point
           .findOneAndUpdate({
             round,
@@ -321,7 +312,6 @@ const resolvers = {
               points: point
             },
             { new: true }).populate('user').populate('round')
-        console.log('check point 2', updatedPoint)
         if (updatedPoint) {
           result.push(updatedPoint)
         } else {
@@ -331,7 +321,6 @@ const resolvers = {
             trackIndex,
             points: point
           })
-          console.log('check point 3')
           try {
             await newPoint
               .save()
@@ -346,7 +335,6 @@ const resolvers = {
           result.push(newPoint)
         }
       }
-      console.log('result', result)
       return result
     },
     addNewTrack: async (root, args, { currentUser }) => {
@@ -381,7 +369,6 @@ const resolvers = {
           })
         }
       }
-      console.log('points', points)
       return await Point
         .find({
           round: args.roundId,
@@ -403,7 +390,6 @@ const resolvers = {
         const points = await Point.find({ trackIndex: maxIndex })
         for (let i = 0; i < points.length; i++) {
           const point = points[i]
-          console.log('point to delete', point)
           await Point.findByIdAndDelete(point._id)
         }
         return maxIndex
@@ -411,7 +397,7 @@ const resolvers = {
       return -1
     },
     deleteUser: async (root, args, { currentUser }) => {
-      if (!currentUser) {
+      if (!currentUser || !currentUser.admin) {
         throw new UserInputError('invalid token');
       }
       const user = await User.findOne({ username: args.username })
@@ -429,35 +415,42 @@ const resolvers = {
       return await User.findByIdAndDelete(user._id)
     },
     deleteAllRounds: async (root, args, { currentUser }) => {
-      if (!currentUser.admin) {
-        throw new UserInputError('Unauthorized deletion of rounds');
+      if (!currentUser || !currentUser.admin) {
+        throw new UserInputError('invalid token');
       }
       await Round.deleteMany({})
       await Point.deleteMany({})
     },
     deleteRound: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new UserInputError('invalid token');
+      }
       const roundId = args.roundId
       const round = await Round.findById(roundId)
       if (!round) {
         throw new UserInputError('Round id not found');
       }
-      if (round.users.filter(user => user === currentUser).length < 1
-        && !currentUser.admin) {
+      if (round.users.filter(user => user === currentUser).length < 1) {
         throw new UserInputError('Unauthorized deletion of round');
       }
       await Point.deleteMany({ round })
       const result = await Round.findByIdAndDelete(roundId)
-      console.log('delete result', result)
       return result
     },
     deleteLocation: async (root, args, { currentUser }) => {
-      return await Location.findOneAndDelete({ name: args.name })
+      if (!currentUser || !currentUser.admin) {
+        throw new UserInputError('invalid token');
+      }
+      const id = args.locationId
+      const res = await Location.findByIdAndDelete(id)
+      return res
     }
   },
+  /*
   Subscription: {
     playAdded: {
       subscribe: () => pubsub.asyncIterator(['PLAY_ADDED'])
     }
-  }
+  }*/
 }
 module.exports = resolvers
